Changes needed to run irods container standalone
I ran into a couple of issues when trying to use the irods container on its own,
rather than as part of a compose.

The main issue was that without reserving more shared memory for the docker container, it wouldn't restart.
This can be seen by a boost error in the rodsServerLog when the server was restarted after the `pkill`.

I also found that it can take some time for the processes to finish, otherwise the restart would fail as the port was still held open.